Rearrange test cases
